+++ src/client/classes/instance.class.js
@@ -85,7 +85,10 @@ export class Instance {
 
         // allocates a unique id for this Tabbed component
         this.#id = 'tabbed-'+Random.id();
-        //console.debug( 'instanciating', this.name(), this.id());
+
+        if( Tabbed.configure().verbosity & Tabbed.C.Verbose.INSTANCIATIONS ){
+            console.debug( 'pwix:tabbed instanciating', this.name(), this.id());
+        }
 
         // setup the parms if they are provided here
         if( Object.keys( opts ) > 1 ){
@@ -421,7 +424,9 @@ export class Instance {
             value = _.isFunction( value ) ? value() : value;
             assert( _.isArray( value ), 'pwix:tabbed.Instance() expects an array, got '+value );
             // make sure each provided tab object has a TABBED definition with a Tab instance
-            //console.debug( this.name(), 'tabs.count', value.length );
+            if( Tabbed.configure().verbosity & Tabbed.C.Verbose.INSTANCIATIONS ){
+                console.debug( 'pwix:tabbed', this.name(), 'tabs.count', value.length );
+            }
             for( let i=0 ; i<value.length ; ++i ){
                 let it = value[i];
                 it.TABBED = it.TABBED || {};